Implemented Page-Switching Functionality
