+++ relangapp/src/components/navBar.css
@@ -3,14 +3,19 @@
     height: 100%;
     background-color: #104051;
 
-    border: none;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 
+    border: none;
+    
     text-align: center;
     text-anchor: middle;
     color: white;
 
     font-size: calc(5px + 1vmax);
     font-weight: bold;
+    text-decoration: none;
 }
 
 .navButton:hover {
