+++ relangapp/src/pages/frontpage.css
@@ -47,7 +47,7 @@
     height: auto;
 
     color: white;
-    font-size: 1.2vmax;
+    font-size: 1.1vmax;
     text-align: left;
 }
 
@@ -113,7 +113,7 @@
     height: auto;
 
     color: white;
-    font-size: 1.2vmax;
+    font-size: 1.15vmax;
     text-align: left;
 }
